toolchain/binutils: Add binutils 2.21.1a
binutils 2.21.1a is the last official binutils release containing
a linker that does not cause ICS to crash on startup.
Unfortunately it still triggers the internal error in
fix_exidx_coverage - making it unsuitable to link parts of ICS
(libbcinfo.so)

But it is a good starting point to debug.

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>